added migration to change sighiting and pairing coords to floats

+++ server/db/migrations/20260104211323_pairings.js
@@ -11,8 +11,8 @@ export function up(knex) {
     table.integer('bird2_id')
     table.string('location')
     table.string('treatment')
-    table.float('nztm_easting')
-    table.float('nztm_northing')
+    table.integer('nztm_easting')
+    table.integer('nztm_northing')
   })
 }
